refactor: 워크플로우 에디터 스토어 위치 정리
