+++ src/features/add-node/model/useAddNode.ts
@@ -5,7 +5,7 @@ import type { Node } from "@xyflow/react";
 import { NODE_REGISTRY } from "@/entities/node";
 import type { DataType } from "@/entities/node";
 import type { FlowNodeData, NodeType } from "@/entities/node";
-import { useWorkflowStore } from "@/shared";
+import { useWorkflowStore } from "@/features/workflow-editor";
 
 const BASE_POSITION = { x: 250, y: 250 };
 const OFFSET_STEP = 30;
