+++ src/features/configure-node/ui/panels/CommunicationPanel.tsx
@@ -5,7 +5,7 @@ import {
   getNodePresentation,
   getTypedConfig,
 } from "@/entities/node";
-import { useWorkflowStore } from "@/shared";
+import { useWorkflowStore } from "@/features/workflow-editor";
 
 import type { NodePanelProps } from "../../model";
