+++ src/features/configure-node/ui/panels/GenericNodePanel.tsx
@@ -1,7 +1,7 @@
 import { Box, Text } from "@chakra-ui/react";
 
 import { getNodePresentation } from "@/entities/node";
-import { useWorkflowStore } from "@/shared";
+import { useWorkflowStore } from "@/features/workflow-editor";
 
 import type { NodePanelProps } from "../../model";
