+++ src/pages/workflow-editor/WorkflowEditorPage.tsx
@@ -7,11 +7,11 @@ import { ReactFlowProvider } from "@xyflow/react";
 
 import { useWorkflowQuery } from "@/entities/workflow";
 import { ServiceSelectionPanel } from "@/features/add-node";
+import { useWorkflowStore } from "@/features/workflow-editor";
 import {
   EDITOR_CANVAS_AREA_ID,
   ROUTE_PATHS,
   hydrateStore,
-  useWorkflowStore,
 } from "@/shared";
 import { Canvas, EditorToolbar, InputPanel, OutputPanel } from "@/widgets";
