+++ src/shared/index.ts
@@ -1,7 +1,6 @@
 export * from "./components";
 export * from "./constants";
 export * from "./libs";
-export * from "./model";
 export * from "./theme";
 export * from "./types";
 export * from "./ui";
